feat(core,legacy): add support for Ethereum 64-bit chain_id
* Changes from original PR

* Now that we are rejecting chain_ids of 0, we need to have the tests set the chain_ids to at least 1.

* Ran 'make gen' and uploaded changed files.

* Ran make style_check and fixed reported errors

* Added changelog files

* Reverted changes concerning chain_id 0 being rejected.

* Adds tests for MAX_CHAIN_ID and MAX_CHAIN_ID+1.  Also reverts MAX_CHAIN_ID to the previous value.

* Added missing whitespace around arithmetic operator.

+++ python/src/trezorlib/messages.py
@@ -4264,7 +4264,7 @@ class EthereumSignTx(protobuf.MessageType):
         6: protobuf.Field("value", "bytes", repeated=False, required=False),
         7: protobuf.Field("data_initial_chunk", "bytes", repeated=False, required=False),
         8: protobuf.Field("data_length", "uint32", repeated=False, required=False),
-        9: protobuf.Field("chain_id", "uint32", repeated=False, required=False),
+        9: protobuf.Field("chain_id", "uint64", repeated=False, required=False),
         10: protobuf.Field("tx_type", "uint32", repeated=False, required=False),
     }
 
@@ -4306,7 +4306,7 @@ class EthereumSignTxEIP1559(protobuf.MessageType):
         7: protobuf.Field("value", "bytes", repeated=False, required=True),
         8: protobuf.Field("data_initial_chunk", "bytes", repeated=False, required=False),
         9: protobuf.Field("data_length", "uint32", repeated=False, required=True),
-        10: protobuf.Field("chain_id", "uint32", repeated=False, required=True),
+        10: protobuf.Field("chain_id", "uint64", repeated=False, required=True),
         11: protobuf.Field("access_list", "EthereumAccessList", repeated=True, required=False),
     }
